Fixed bug: SemaphoreNeighborInject transformations were implemented ambiguously (some tools used route-semaphore matches, some used route matches)

+++ trainbenchmark-tool-neo4j/src/main/java/hu/bme/mit/trainbenchmark/benchmark/neo4j/transformations/core/inject/Neo4jCoreTransformationInjectSemaphoreNeighbor.java
@@ -31,7 +31,9 @@ public class Neo4jCoreTransformationInjectSemaphoreNeighbor extends Neo4jCoreTra
 		for (final Neo4jSemaphoreNeighborInjectMatch match : matches) {
 			final Iterable<Relationship> entries = match.getRoute().getRelationships(Neo4jConstants.relationshipTypeEntry);
 			for (final Relationship entry : entries) {
-				entry.delete();
+				if (entry.getEndNode().equals(match.getSemaphore())) {
+					entry.delete();
+				}
 			}
 		}
 	}
